moving bmc to a separate file
