+++ emukit/quadrature/loop/__init__.py
@@ -2,5 +2,5 @@
 # SPDX-License-Identifier: Apache-2.0
 
 
-from .quadrature_loop import VanillaBayesianQuadratureLoop  # noqa: F401
-from .quadrature_loop import SimpleBayesianMonteCarlo  # noqa: F401
+from .vanilla_bq_loop import VanillaBayesianQuadratureLoop  # noqa: F401
+from .bayesian_monte_carlo_loop import SimpleBayesianMonteCarlo  # noqa: F401
