+++ integration_tests/emukit/quadrature/test_bayesian_monte_carlo.py
@@ -1,4 +1,4 @@
-# Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
 
@@ -6,7 +6,7 @@ import numpy as np
 import GPy
 
 from emukit.quadrature.methods.vanilla_bq import VanillaBayesianQuadrature
-from emukit.quadrature.loop.vanilla_bq_loop import VanillaBayesianQuadratureLoop
+from emukit.quadrature.loop.bayesian_monte_carlo_loop import SimpleBayesianMonteCarlo
 from emukit.core.loop.user_function import UserFunctionWrapper
 from emukit.model_wrappers.gpy_quadrature_wrappers import QuadratureRBFLebesgueMeasure, RBFGPy, BaseGaussianProcessGPy
 
@@ -17,7 +17,7 @@ def func(x):
     return np.ones((x.shape[0], 1))
 
 
-def test_vanilla_bq_loop():
+def test_bayesian_monte_carlo_loop():
     init_size = 5
     x_init = np.random.rand(init_size, 2)
     y_init = np.random.rand(init_size, 1)
@@ -28,7 +28,7 @@ def test_vanilla_bq_loop():
     emukit_qrbf = QuadratureRBFLebesgueMeasure(RBFGPy(gpy_model.kern), integral_bounds=bounds)
     emukit_model = BaseGaussianProcessGPy(kern=emukit_qrbf, gpy_model=gpy_model)
     emukit_method = VanillaBayesianQuadrature(base_gp=emukit_model, X=x_init, Y=y_init)
-    emukit_loop = VanillaBayesianQuadratureLoop(model=emukit_method)
+    emukit_loop = SimpleBayesianMonteCarlo(model=emukit_method)
 
     num_iter = 5
     emukit_loop.run_loop(user_function=UserFunctionWrapper(func), stopping_condition=num_iter)
@@ -37,7 +37,7 @@ def test_vanilla_bq_loop():
     assert emukit_loop.loop_state.Y.shape[0] == num_iter + init_size
 
 
-def test_vanilla_bq_loop_initial_state():
+def test_bayesian_monte_carlo_loop_initial_state():
     x_init = np.random.rand(5, 2)
     y_init = np.random.rand(5, 1)
 
@@ -48,7 +48,7 @@ def test_vanilla_bq_loop_initial_state():
     emukit_qrbf = QuadratureRBFLebesgueMeasure(RBFGPy(gpy_model.kern), integral_bounds=bounds)
     emukit_model = BaseGaussianProcessGPy(kern=emukit_qrbf, gpy_model=gpy_model)
     emukit_method = VanillaBayesianQuadrature(base_gp=emukit_model, X=x_init, Y=y_init)
-    emukit_loop = VanillaBayesianQuadratureLoop(model=emukit_method)
+    emukit_loop = SimpleBayesianMonteCarlo(model=emukit_method)
 
     assert_array_equal(emukit_loop.loop_state.X, x_init)
     assert_array_equal(emukit_loop.loop_state.Y, y_init)
